Trait divrem::DivFloor

pub trait DivFloor<RHS = Self>: Div<RHS> {
    // Required method
    fn div_floor(self, other: RHS) -> <Self as Div<RHS>>::Output;
}

Floored division.

Floors the quotient and effectively rounds towards negative infinity.

This is equivalent to the div function in Haskell.

Required Methods

fn div_floor(self, other: RHS) -> <Self as Div<RHS>>::Output

Performs the floored division operation.