Trait divrem::DivRemCeil

pub trait DivRemCeil<RHS = Self>: DivRem<RHS> {
    // Required method
    fn div_rem_ceil(self, other: RHS) -> <Self as DivRem<RHS>>::Output;
}

Ceiled division and remainder.

Ceils the quotient and effectively rounds towards positive infinity. The sign of the modulus is always the opposite of the sign of the divisor or zero.

Required Methods

fn div_rem_ceil(self, other: RHS) -> <Self as DivRem<RHS>>::Output

Performs the ceiled division operation with remainder.