Trait divrem::DivRemFloor

pub trait DivRemFloor<RHS = Self>: DivRem<RHS> {
    fn div_rem_floor(self, other: RHS) -> Self::Output;
}

Floored division and remainder.

Floors the quotient and effectively rounds towards negative infinity. The sign of the modulus is always the same as the sign of the divisor or zero.

This is equivalent to the divMod function in Haskell.

Required Methods

fn div_rem_floor(self, other: RHS) -> Self::Output

Performs the floored division operation with remainder.