Trait divrem::RemCeil

pub trait RemCeil<RHS = Self>: Rem<RHS> {
    fn rem_ceil(self, other: RHS) -> Self::Output;
}

Ceiled division remainder.

The sign of the modulus is always the opposite of the sign of the divisor or zero.

Required Methods

fn rem_ceil(self, other: RHS) -> Self::Output

Returns the remainder of the ceiled division operation.